requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-filter
drf-spectacular